azure_agent/requirements.txt
streamlit
langchain
langchain-community
azure-search-documents
azure-identity
openai
python-dotenv
tiktoken
